src/requirements.txt
PyYAML==6.0.1
pytest==8.3.4
